Update to handlespace management with Internet-32-checksum. Update to UDP-like registrar connection handling including incoming registrar change after takeover. Fractal example programs now correctly uses host-independent double datatype. Bugfixes in fractal PU.
git-svn-id: https://svn.tdr.iem.uni-due.de:8443/svn/sctp/trunk/rsplib2@662 516812f1-d3fd-0310-9ea4-9aededebcb87

+++ rsplib/componentstatusprotocol.c
@@ -102,7 +102,7 @@ void componentStatusGetComponentAddress(char*        componentAddress,
       addresses = gatherLocalAddresses(&addressArray);
    }
    if(addresses > 0) {
-      for(i = 0;i < addresses;i++) {
+      for(i = 0;i < (size_t)addresses;i++) {
          if(getScope((const struct sockaddr*)&addressArray[i]) >= 6) {
             if(address2string((const struct sockaddr*)&addressArray[i], (char*)&str, sizeof(str), false)) {
                if(componentAddress[0] != 0x00) {
